agents/strands-agents/strands-aws-cost-optimization/main.py
"""

"""

from aws_billing_management_agent import aws_billing_management_agent
from strands import Agent
from strands.models import BedrockModel
from strands_tools import think
from bedrock_agentcore.runtime import BedrockAgentCoreApp


app = BedrockAgentCoreApp()

bedrock_model = BedrockModel(model_id="us.anthropic.claude-3-7-sonnet-20250219-v1:0")

SUPERVISOR_AGENT_PROMPT = """You are an AWS Cost Optimization Router Agent, a sophisticated orchestrator designed to coordinate AWS billing and cost management operations. Your role is to:

1. Analyze the user's request and determine the appropriate tool to use.
2. Execute the selected tool and provide comprehensive cost optimization guidance.
3. Provide actionable recommendations with specific savings opportunities.

The available tools are:
- aws_billing_management_agent: For AWS cost analysis, billing management, optimization recommendations, and financial operations.
- think: For general reasoning and analysis when direct AWS billing tools aren't needed.

You will use the tools in the following priority order:
1. If the user's request is related to AWS costs, billing, budgets, reservations, savings plans, cost optimization, or financial analysis, use the aws_billing_management_agent tool.
2. For general questions or non-financial AWS topics, use the think tool.

Key areas handled by aws_billing_management_agent:
- Cost analysis and forecasting
- Budget management and monitoring
- Reserved Instances and Savings Plans optimization
- Cost anomaly detection
- Service-specific cost optimization (EC2, RDS, Lambda, S3, etc.)
- Compute Optimizer recommendations
- Storage cost optimization
- Pricing analysis and comparisons

Always prioritize providing data-driven, actionable cost optimization guidance with specific dollar amounts and implementation steps.
"""

supervisor_agent = Agent(
    system_prompt=SUPERVISOR_AGENT_PROMPT,
    model=bedrock_model,
    tools=[aws_billing_management_agent, think],
)


@app.entrypoint
def strands_agent_bedrock(payload):
    """
    Invoke the agent with a payload
    """
    user_input = payload.get("prompt")
    print("User input:", user_input)
    response = supervisor_agent(user_input)
    return response.message['content'][0]['text']


# Example usage
if __name__ == "__main__":
    app.run()